+++ contracts/contract/src/fungible_token.rs
@@ -94,15 +94,21 @@ impl FungibleTokenReceiver for Contract {
                 }
                 TokenReceiverMsg::SwapReference { swap_ref } => {
                     let config = self.internal_config();
-                    assert!(sender_id == config.ref_exchange_id || sender_id == config.dcl_id.expect("Missing dcl id"), "Not allow");
+                    let mut account = self.internal_unwrap_margin_account(&swap_ref.account_id);
+                    let action_ts = account.position_latest_actions.remove(&swap_ref.pos_id).expect("There is no action for the position").0;
+                    if sender_id == config.owner_id {
+                        require!(env::block_timestamp() - action_ts >= sec_to_nano(self.internal_margin_config().max_position_action_wait_sec), "Please wait for the position action");
+                    } else {
+                        require!(sender_id == config.ref_exchange_id || sender_id == config.dcl_id.expect("Missing dcl id"), "Not allow");
+                    }
                     if swap_ref.op == "open" {
-                        self.on_open_trade_return(&swap_ref.account_id, amount, &swap_ref);
+                        self.on_open_trade_return(account, amount, &swap_ref);
                     } else if swap_ref.op == "decrease"
                         || swap_ref.op == "close"
                         || swap_ref.op == "liquidate"
                         || swap_ref.op == "forceclose"
                     {
-                        let event = self.on_decrease_trade_return(&swap_ref.account_id, amount, &swap_ref);
+                        let event = self.on_decrease_trade_return(account, amount, &swap_ref);
                         events::emit::margin_decrease_succeeded(&swap_ref.op, event);
                     }
                     return PromiseOrValue::Value(U128(0));