+++ releases/release_notes.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+Version 0.15.0
+```bash=
+# codehash: 3kn9awtt4caZJXmCZ6WqZVhXTUqmWDPLREU1Kg3vxCCQ
+```
+- Prohibit update_config from modifying owner_id, add a set_owner_id function that only allows setting a registered user as the owner.
+- Allow the owner to initiate SwapReference if the position action times out and is not completed.
+- If the liquidator deregisters their account during the liquidation process, all liquidation profits will go to the owner.
+
 Version 0.14.3
 ```bash=
 # codehash: TdQAXwbxGsnwe2icachwgDBVbAnpQNh1ZogoTT15GdF